+++ domain/src/main/java/tachiyomi/domain/history/model/HistorySourceFilter.kt
@@ -1,0 +1,6 @@
+package tachiyomi.domain.history.model
+
+sealed interface HistorySourceFilter {
+    data class Library(val excludedLocalSourceId: Long) : HistorySourceFilter
+    data class Local(val localSourceId: Long) : HistorySourceFilter
+}